Split productFactory
